refactor(pci-instances): restructure types folder
